ER-1408 - moving AlertViewModelService to shared
