+++ src/Employer/Employer.Web/Views/Dashboard/_EmployerRevokedTransferredVacanciesAlert.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Esfa.Recruit.Vacancies.Client.Domain.Entities
-@model Esfa.Recruit.Employer.Web.ViewModels.Dashboard.TransferredVacanciesAlertViewModel
+@model Esfa.Recruit.Shared.Web.ViewModels.Alerts.TransferredVacanciesAlertViewModel
 
 <div class="govuk-grid-row">
     <div class="govuk-grid-column-full">
