+++ src/Employer/Employer.Web/Views/Dashboard/_WithdrawnVacanciesAlert.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Esfa.Recruit.Vacancies.Client.Domain.Entities
-@model Esfa.Recruit.Employer.Web.ViewModels.Dashboard.WithdrawnVacanciesAlertViewModel
+@model Esfa.Recruit.Shared.Web.ViewModels.Alerts.WithdrawnVacanciesAlertViewModel
 
 <div class="govuk-grid-row">
     <div class="govuk-grid-column-full">
