+++ src/Shared/Recruit.Shared.Web/Recruit.Shared.Web.csproj
@@ -4,6 +4,7 @@
     <RootNamespace>Esfa.Recruit.Shared.Web</RootNamespace>
   </PropertyGroup>
   <ItemGroup>
+    <PackageReference Include="Humanizer.Core.uk" Version="2.6.2" />
     <PackageReference Include="Microsoft.AspNetCore.Diagnostics" Version="2.1.1" />
     <PackageReference Include="Microsoft.AspNetCore.Mvc" Version="2.1.3" />
     <PackageReference Include="NWebsec.AspNetCore.Mvc.TagHelpers" Version="2.0.0" />
