+++ src/Shared/Recruit.Shared.Web/ViewModels/Alerts/TransferredVacanciesAlertViewModel.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using Esfa.Recruit.Shared.Web.Extensions;
 using Humanizer;
 
-namespace Esfa.Recruit.Employer.Web.ViewModels.Dashboard
+namespace Esfa.Recruit.Shared.Web.ViewModels.Alerts
 {
     public class TransferredVacanciesAlertViewModel
     {
