+++ src/Shared/Recruit.Shared.Web/ViewModels/Alerts/WithdrawnVacanciesAlertViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace Esfa.Recruit.Employer.Web.ViewModels.Dashboard
+namespace Esfa.Recruit.Shared.Web.ViewModels.Alerts
 {
     public class WithdrawnVacanciesAlertViewModel
     {
