+++ src/Shared/UnitTests/Shared.Web/Services/AlertViewModelServiceTests/AlertViewModelServiceTestBase.cs
@@ -1,7 +1,7 @@
 ﻿
 using System;
 
-namespace Esfa.Recruit.Employer.UnitTests.Employer.Web.Services.AlertViewModelServiceTests
+namespace Esfa.Recruit.Vacancies.Client.UnitTests.Shared.Web.Services.AlertViewModelServiceTests
 {
     public abstract class AlertViewModelServiceTestBase
     {
